+++ views/checklists/index.hbs
@@ -75,4 +75,4 @@
 
 <script src="/javascripts/todo.js"></script>
 <script src="/javascripts/checklist-client.js"></script>
-{{!-- <script src="/javascripts/socket-todo.js"></script> --}}
+<script src="/javascripts/todo-client.js"></script>
